requirements.txt
playwright
pandas
openai>=1.50.0
python-docx>=1.1.2
python-dotenv>=1.0.1
openpyxl
beautifulsoup4
